Implement Object have Card Game Logic
