+++ CardGameApp/CardGameApp/GameParticipates/Dealer.swift
@@ -63,45 +63,5 @@ class Dealer : GameParticipate {
             return false
         }
     }
-    
-    func judgeLogic(to players: Players) {
-        players.judgePlayersState { playerCards in
-            let tempCard = cards.sorted(by: {$0.number.rawValue < $1.number.rawValue})
-            guard judgeFourCard(of: tempCard) else { return .fourCard }
-            guard judgeTripple(of: tempCard) else { return .tripple }
-            guard judgeTwoPair(of: tempCard) else { return .twoPair }
-            guard judgeOnePair(of: tempCard) else { return .onePair }
-            return .noPair
-        }
-    }
-    
-    private func judgeFourCard(of sortedCard: [Card]) -> Bool {
-        for index in 0..<sortedCard.count-3 {
-            if sortedCard[index].number == sortedCard[index+1].number && sortedCard[index+1].number == sortedCard[index+2].number && sortedCard[index+2].number == sortedCard[index+3].number { return true }
-        }
-        return false
-    }
-    
-    private func judgeTripple(of sortedCard: [Card]) -> Bool {
-        for index in 0..<sortedCard.count-2 {
-            if sortedCard[index].number == sortedCard[index+1].number && sortedCard[index+1].number == sortedCard[index+2].number { return true }
-        }
-        return false
-    }
-    
-    private func judgeTwoPair(of sortedCard: [Card]) -> Bool {
-        var pairCount: Int = 0
-        for index in 0..<sortedCard.count-1 {
-            if sortedCard[index].number == sortedCard[index+1].number { pairCount += 1 }
-        }
-        if pairCount > 1 { return true }
-        return false
-    }
-    
-    private func judgeOnePair(of sortedCard: [Card]) -> Bool {
-        for index in 0..<sortedCard.count-1 {
-            if sortedCard[index].number == sortedCard[index+1].number { return true }
-        }
-        return false
-    }
+
 }
